+++ src/ast/nodes/LogicalExpression.ts
@@ -8,6 +8,10 @@ import {
 } from './shared/Expression';
 import { NodeType } from './NodeType';
 import { ExpressionNode, NodeBase } from './shared/Node';
+import { NodeRenderOptions, RenderOptions } from '../../utils/renderHelpers';
+import MagicString from 'magic-string';
+import { BLANK } from '../../utils/blank';
+import CallExpression from './CallExpression';
 
 export type LogicalOperator = '||' | '&&';
 
@@ -17,17 +21,13 @@ export default class LogicalExpression extends NodeBase {
 	left: ExpressionNode;
 	right: ExpressionNode;
 
-	reassignPath(path: ObjectPath, options: ExecutionPathOptions) {
-		path.length > 0 && this._forEachRelevantBranch(node => node.reassignPath(path, options));
-	}
-
 	forEachReturnExpressionWhenCalledAtPath(
 		path: ObjectPath,
 		callOptions: CallOptions,
 		callback: ForEachReturnExpressionCallback,
 		options: ExecutionPathOptions
 	) {
-		this._forEachRelevantBranch(node =>
+		this.forEachRelevantBranch(node =>
 			node.forEachReturnExpressionWhenCalledAtPath(path, callOptions, callback, options)
 		);
 	}
@@ -35,34 +35,30 @@ export default class LogicalExpression extends NodeBase {
 	getValue(): any {
 		const leftValue = this.left.getValue();
 		if (leftValue === UNKNOWN_VALUE) return UNKNOWN_VALUE;
-		if ((leftValue && this.operator === '||') || (!leftValue && this.operator === '&&')) {
-			return leftValue;
-		}
+		if (!leftValue === (this.operator === '&&')) return leftValue;
 		return this.right.getValue();
 	}
 
 	hasEffects(options: ExecutionPathOptions): boolean {
+		if (this.left.hasEffects(options)) return true;
 		const leftValue = this.left.getValue();
 		return (
-			this.left.hasEffects(options) ||
-			((leftValue === UNKNOWN_VALUE ||
-				(!leftValue && this.operator === '||') ||
-				(leftValue && this.operator === '&&')) &&
-				this.right.hasEffects(options))
+			(leftValue === UNKNOWN_VALUE || !leftValue === (this.operator === '||')) &&
+			this.right.hasEffects(options)
 		);
 	}
 
 	hasEffectsWhenAccessedAtPath(path: ObjectPath, options: ExecutionPathOptions): boolean {
 		return (
 			path.length > 0 &&
-			this._someRelevantBranch(node => node.hasEffectsWhenAccessedAtPath(path, options))
+			this.someRelevantBranch(node => node.hasEffectsWhenAccessedAtPath(path, options))
 		);
 	}
 
 	hasEffectsWhenAssignedAtPath(path: ObjectPath, options: ExecutionPathOptions): boolean {
 		return (
 			path.length === 0 ||
-			this._someRelevantBranch(node => node.hasEffectsWhenAssignedAtPath(path, options))
+			this.someRelevantBranch(node => node.hasEffectsWhenAssignedAtPath(path, options))
 		);
 	}
 
@@ -71,9 +67,52 @@ export default class LogicalExpression extends NodeBase {
 		callOptions: CallOptions,
 		options: ExecutionPathOptions
 	): boolean {
-		return this._someRelevantBranch(node =>
+		return this.someRelevantBranch(node =>
 			node.hasEffectsWhenCalledAtPath(path, callOptions, options)
 		);
+	}
+
+	includeInBundle() {
+		let addedNewNodes = !this.included;
+		this.included = true;
+		const leftValue = this.left.getValue();
+		if (
+			(leftValue === UNKNOWN_VALUE ||
+				!leftValue === (this.operator === '&&') ||
+				this.left.shouldBeIncluded()) &&
+			this.left.includeInBundle()
+		)
+			addedNewNodes = true;
+		if (
+			(leftValue === UNKNOWN_VALUE || !leftValue === (this.operator === '||')) &&
+			this.right.includeInBundle()
+		)
+			addedNewNodes = true;
+		return addedNewNodes;
+	}
+
+	reassignPath(path: ObjectPath, options: ExecutionPathOptions) {
+		if (path.length > 0) this.forEachRelevantBranch(node => node.reassignPath(path, options));
+	}
+
+	render(
+		code: MagicString,
+		options: RenderOptions,
+		{ renderedParentType, isCalleeOfRenderedParent }: NodeRenderOptions = BLANK
+	) {
+		if (!this.module.graph.treeshake || (this.left.included && this.right.included)) {
+			super.render(code, options);
+		} else {
+			const branchToRetain = this.left.included ? this.left : this.right;
+			code.remove(this.start, branchToRetain.start);
+			code.remove(branchToRetain.end, this.end);
+			branchToRetain.render(code, options, {
+				renderedParentType: renderedParentType || this.parent.type,
+				isCalleeOfRenderedParent: renderedParentType
+					? isCalleeOfRenderedParent
+					: (<CallExpression>this.parent).callee === this
+			});
+		}
 	}
 
 	someReturnExpressionWhenCalledAtPath(
@@ -82,31 +121,30 @@ export default class LogicalExpression extends NodeBase {
 		predicateFunction: SomeReturnExpressionCallback,
 		options: ExecutionPathOptions
 	): boolean {
-		return this._someRelevantBranch(node =>
+		return this.someRelevantBranch(node =>
 			node.someReturnExpressionWhenCalledAtPath(path, callOptions, predicateFunction, options)
 		);
 	}
 
-	_forEachRelevantBranch(callback: (node: ExpressionNode) => void) {
+	private forEachRelevantBranch(callback: (node: ExpressionNode) => void) {
 		const leftValue = this.left.getValue();
 		if (leftValue === UNKNOWN_VALUE) {
 			callback(this.left);
 			callback(this.right);
-		} else if ((leftValue && this.operator === '||') || (!leftValue && this.operator === '&&')) {
+		} else if (!leftValue === (this.operator === '&&')) {
 			callback(this.left);
 		} else {
 			callback(this.right);
 		}
 	}
 
-	_someRelevantBranch(predicateFunction: PredicateFunction) {
+	private someRelevantBranch(predicateFunction: PredicateFunction) {
 		const leftValue = this.left.getValue();
 		if (leftValue === UNKNOWN_VALUE) {
 			return predicateFunction(this.left) || predicateFunction(this.right);
 		}
-		if ((leftValue && this.operator === '||') || (!leftValue && this.operator === '&&')) {
-			return predicateFunction(this.left);
-		}
-		return predicateFunction(this.right);
+		return !leftValue === (this.operator === '&&')
+			? predicateFunction(this.left)
+			: predicateFunction(this.right);
 	}
 }