+++ test/form/samples/conditional-put-parens-around-sequence/_expected/umd.js
@@ -4,11 +4,11 @@
 	(factory());
 }(this, (function () { 'use strict';
 
-	var a = (foo(), 3);
-	var b = (bar(), 6);
+	var a = (foo(), 3 );
+	var b = (bar(), 6 );
 	foo( a, b );
 
 	// verify works with no whitespace
-	bar((foo(), 2),(bar(), 8));
+	bar((foo(),2),(bar(),8));
 
 })));