+++ test/form/samples/mutate-logical-expression/_expected/system.js
@@ -8,11 +8,11 @@ System.register([], function (exports, module) {
 			logicalAExp.bar = 1;
 
 			var bExp = exports('bExp', {});
-			var logicalBExp = false || bExp;
+			var logicalBExp = bExp;
 			logicalBExp.bar = 1;
 
 			var cExp = exports('cExp', {});
-			var logicalCExp = true && cExp;
+			var logicalCExp = cExp;
 			logicalCExp.bar = 1;
 
 		}