+++ test/form/samples/side-effects-logical-expressions/_expected/system.js
@@ -4,8 +4,8 @@ System.register([], function (exports, module) {
 		execute: function () {
 
 			// effect
-			false || console.log( 'effect' );
-			true && console.log( 'effect' );
+			console.log( 'effect' );
+			console.log( 'effect' );
 			console.log( 'effect' ) || {};
 			console.log( 'effect' ) && {};
 
@@ -17,24 +17,24 @@ System.register([], function (exports, module) {
 			};
 
 			// effect
-			(false || foo).effect;
-			(true && foo).effect;
+			(foo).effect;
+			(foo).effect;
 
 			// effect
-			(false || null).foo = 1;
-			(true && null).foo = 1;
+			(null).foo = 1;
+			(null).foo = 1;
 
 			// effect
-			(true || (() => {}))();
-			(false && (() => {}))();
-			(false || (() => console.log( 'effect' )))();
-			(true && (() => console.log( 'effect' )))();
+			(true)();
+			(false)();
+			(() => console.log( 'effect' ))();
+			(() => console.log( 'effect' ))();
 
 			// effect
-			(true || (() => () => {}))()();
-			(false && (() => () => {}))()();
-			(false || (() => () => console.log( 'effect' )))()();
-			(true && (() => () => console.log( 'effect' )))()();
+			(true)()();
+			(false)()();
+			(() => () => console.log( 'effect' ))()();
+			(() => () => console.log( 'effect' ))()();
 
 		}
 	};