+++ test/form/samples/tree-shake-logical-expressions/_expected/amd.js
@@ -1,0 +1,19 @@
+define(function () { 'use strict';
+
+	function getStringA() {
+		return 'A';
+	}
+
+	console.log(getStringA());
+
+	console.log(false);
+
+	console.log(true);
+
+	function getStringD() {
+		return 'D';
+	}
+
+	console.log(getStringD());
+
+});