+++ test/form/samples/wrap-simplified-expressions/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'wraps simplified expressions that have become callees if necessary'
+};