+++ test/function/samples/side-effects-only-default-exports/_config.js
@@ -1,0 +1,18 @@
+const assert = require('assert');
+const mutated = {};
+
+module.exports = {
+	description: 'Wraps inlined default exports which are rendered for side-effects only',
+	options: {
+		external: ['external']
+	},
+	context: {
+		require: () => mutated
+	},
+	bundle() {
+		assert.ok(mutated.mutated);
+		assert.ok(mutated.veryMutated);
+		assert.ok(mutated.extremelyMutated);
+		assert.ok(mutated.utterlyMutated);
+	}
+};