+++ test/function/samples/side-effects-only-default-exports/bar.js
@@ -1,0 +1,3 @@
+import x from 'external';
+
+export default (x.veryMutated = true, x.extremelyMutated = true);