+++ test/function/samples/side-effects-only-default-exports/baz.js
@@ -1,0 +1,6 @@
+import x from 'external';
+
+export default true ? {
+	a: x.totallyMutated = true,
+	makeSureThisIsAnObject: true
+} : null;