+++ test/function/samples/side-effects-only-default-exports/foo.js
@@ -1,0 +1,6 @@
+import x from 'external';
+
+export default {
+	a: x.mutated = true,
+	makeSureThisIsAnObject: true
+};