+++ test/function/samples/side-effects-only-default-exports/main.js
@@ -1,0 +1,4 @@
+import foo from './foo.js';
+import bar from './bar.js';
+import baz from './baz.js';
+import quux from './quux.js';