+++ test/function/samples/wraps-object-expressions-as-statements/main.js
@@ -1,0 +1,10 @@
+const foo = {};
+
+// new parent is expression statement
+true && {
+	a: foo.a = true,
+	b: foo.a = true
+};
+assert.ok(foo.a);
+
+export default foo;